設計に関する問題を修正
